Added easy extension of custom merge results. Example fib merge rules added. Code altered to make use of this new style

+++ src/engine.h
@@ -13,8 +13,8 @@
 
 struct gamestate {
     /* Game state */
-    long *grid_data_ptr;
-    long **grid;
+    int *grid_data_ptr;
+    int **grid;
     int gridsize;
     int moved;
     long score;
